+++ src/stirling/vulkan/index_buffer.h
@@ -15,7 +15,7 @@ namespace stirling {
 	namespace vulkan {
 		class IndexBuffer : public Buffer {
 		public:
-			IndexBuffer(const Device& device, const std::vector<uint16_t>& indices);
+			IndexBuffer(const Device& device, const std::vector<uint32_t>& indices);
 
 			int size() const;
